feat: uncaught ex, crashed=true, handled=false (#97)

+++ sentry-core/src/main/java/io/sentry/core/protocol/SentryException.java
@@ -8,7 +8,7 @@ public final class SentryException implements IUnknownPropertiesConsumer {
   private String type;
   private String value;
   private String module;
-  private Integer threadId;
+  private Long threadId;
   private SentryStackTrace stacktrace;
   private Mechanism mechanism;
 
@@ -70,7 +70,7 @@ public final class SentryException implements IUnknownPropertiesConsumer {
    *
    * @return the thread id.
    */
-  public Integer getThreadId() {
+  public Long getThreadId() {
     return threadId;
   }
 
@@ -79,7 +79,7 @@ public final class SentryException implements IUnknownPropertiesConsumer {
    *
    * @param threadId the thread id.
    */
-  public void setThreadId(Integer threadId) {
+  public void setThreadId(Long threadId) {
     this.threadId = threadId;
   }
